Minor Seller API Changes.

+++ src/main/java/com/bootcamp/ECommerceApplication/entity/ProductVariation.java
@@ -18,7 +18,7 @@ public class ProductVariation {
     private Product product;
 
     private Integer quantityAvailable;
-    private Double price;
+    private Integer price;
     private String metadata;
     private boolean isActive;
     @Transient
@@ -54,11 +54,11 @@ public class ProductVariation {
         this.quantityAvailable = quantityAvailable;
     }
 
-    public Double getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
